markdownSelector and markdownPreview selector tests
