+++ src/components/markdown/Preview.jsx
@@ -1,11 +1,11 @@
 import React from 'react';
-import marked from 'marked';
 import styles from 'github-markdown-css/github-markdown.css';
 import { useSelector } from '../../hooks/appContext';
-import getMarkdown from '../../selectors/getMarkdown';
+import { getMarkdownPreview } from '../../selectors/markdownSelectors';
+
 const Preview = () => {
-  const markdown = useSelector(getMarkdown);
-  const __html = marked(markdown);
+
+  const __html = useSelector(getMarkdownPreview);
   return <div className={styles['markdown-body']} dangerouslySetInnerHTML={{ __html }}></div>;
 };
