--- src/selectors/getMarkdown.js
@@ -1,2 +1,0 @@
-const getMarkdown = state => state.markdown;
-export default getMarkdown;
