+++ src/selectors/markdownSelectors.js
@@ -1,0 +1,5 @@
+import marked from 'marked';
+
+export const getMarkdown = state => state.markdown;
+
+export const getMarkdownPreview = state => marked(getMarkdown(state));
